+++ cli/crates/cli/src/cli_input/federated_graph.rs
@@ -1,0 +1,13 @@
+use clap::Parser;
+
+#[derive(Debug, Parser)]
+pub enum FederatedGraphCommandKind {
+    /// Fetch the GraphQL schema of the federated graph
+    Fetch,
+}
+
+#[derive(Debug, Parser)]
+struct FederatedGraphCommand {
+    #[command(subcommand)]
+    kind: FederatedGraphCommandKind,
+}
