+++ cli/crates/cli/src/cli_input/subgraphs.rs
@@ -1,0 +1,9 @@
+use super::ProjectRef;
+use clap::Parser;
+
+/// List published subgraphs
+#[derive(Debug, Parser)]
+pub struct SubgraphsCommand {
+    #[arg(help = ProjectRef::ARG_DESCRIPTION)]
+    pub project_ref: ProjectRef,
+}
