+++ cli/crates/federated-dev/src/dev/admin.rs
@@ -1,6 +1,6 @@
 mod types;
 
-pub(crate) use self::types::{Header, PublishSubgraphInput, PublishSubgraphSuccess};
+pub(crate) use self::types::{Header, PublishSubgraphInput};
 use super::bus::AdminBus;
 pub(crate) struct MutationRoot;
 
@@ -8,11 +8,7 @@ use async_graphql::{Context, Error, Object};
 
 #[Object]
 impl MutationRoot {
-    pub(crate) async fn publish_subgraph(
-        &self,
-        ctx: &Context<'_>,
-        input: PublishSubgraphInput,
-    ) -> Result<PublishSubgraphSuccess, Error> {
+    pub(crate) async fn publish_subgraph(&self, ctx: &Context<'_>, input: PublishSubgraphInput) -> Result<bool, Error> {
         let bus = ctx.data::<AdminBus>().expect("must be a bus");
         let schema = bus
             .introspect_schema(&input.name, input.url.clone(), input.headers.clone())
@@ -20,7 +16,7 @@ impl MutationRoot {
 
         bus.compose_graph(input.name, input.url, input.headers, schema).await?;
 
-        Ok(PublishSubgraphSuccess::default())
+        Ok(true)
     }
 }
