--- cli/crates/federated-dev/src/dev/composer/messages.rs
@@ -1,1 +1,0 @@
-
